Renamed RecipeBook module to recipe to have more consistent business logic naming
