+++ src/index.ts
@@ -2,4 +2,4 @@ export { Home, RecipeForm, RecipeView } from "./ui";
 export { Stack } from "./Stack";
 export { AppContext } from "./AppContext";
 
-export type { RecipeBook } from "./core/RecipeBook";
+export type { RecipeBook } from "./core/recipe";
